requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-pptx==0.6.21
pydantic==2.5.0

# Database
sqlmodel==0.0.14
sqlalchemy==2.0.23
aiosqlite==0.19.0
greenlet==3.0.1

# Caching
cachetools==5.3.2

# LLM and content generation
openai==1.3.7
python-dotenv==1.0.0

# HTTP client for API calls
httpx==0.25.2

# Data handling
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 